Assignment No. 14
A double-ended queue (deque) is a linear list in which additions and deletions
may be made at either end. Obtain a data representation mapping a deque into a
one- dimensional array. Write C++ program to simulate deque with functions to
add and delete elements from either end of the deque.